+++ R/do_ViolinPlot.R
@@ -7,7 +7,6 @@
 #' @param line_width \strong{\code{\link[base]{numeric}}} | Width of the lines drawn in the plot. Defaults to 1.
 #' @param boxplot_width \strong{\code{\link[base]{numeric}}} | Width of the boxplots. Defaults to 0.2.
 #' @param share.y.lims \strong{\code{\link[base]{logical}}} | When querying multiple features, force the Y axis of all of them to be on the same range of values (this being the max and min of all features combined).
-#' @param legend.ncol \strong{\code{\link[base]{numeric}}} | Number of columns in the legend.
 
 
 #' @return A ggplot2 object containing a Violin Plot.
@@ -42,7 +41,9 @@ do_ViolinPlot <- function(sample,
                           ncol = NULL,
                           share.y.lims = FALSE,
                           legend.title = NULL,
-                          legend.ncol = NULL){
+                          legend.ncol = NULL,
+                          legend.nrow = NULL,
+                          legend.byrow = FALSE){
   check_suggests(function_name = "do_ViolinPlot")
   # Check if the sample provided is a Seurat object.
   check_Seurat(sample = sample)
@@ -57,7 +58,8 @@ do_ViolinPlot <- function(sample,
   logical_list <- list("plot_boxplot" = plot_boxplot,
                        "plot.grid" = plot.grid,
                        "flip" = flip,
-                       "share.y.lims" = share.y.lims)
+                       "share.y.lims" = share.y.lims,
+                       "legend.byrow" = legend.byrow)
   check_type(parameters = logical_list, required_type = "logical", test_function = is.logical)
   # Check numeric parameters.
   numeric_list <- list("pt.size" = pt.size,
@@ -67,7 +69,8 @@ do_ViolinPlot <- function(sample,
                        "boxplot_width" = boxplot_width,
                        "rotate_x_axis_labels" = rotate_x_axis_labels,
                        "ncol" = ncol,
-                       "legend.ncol" = legend.ncol)
+                       "legend.ncol" = legend.ncol,
+                       "legend.nrow" = legend.nrow)
   check_type(parameters = numeric_list, required_type = "numeric", test_function = is.numeric)
   # Check character parameters.
   character_list <- list("legend.position" = legend.position,
@@ -181,6 +184,8 @@ do_ViolinPlot <- function(sample,
                        caption = plot.caption) +
          ggplot2::guides(fill = ggplot2::guide_legend(title = legend.title,
                                                       ncol = legend.ncol,
+                                                      nrow = legend.nrow,
+                                                      byrow = legend.byrow,
                                                       title.position = "top")) +
          ggplot2::theme_minimal(base_size = font.size) +
          ggplot2::theme(axis.text.x = ggplot2::element_text(color = "black",